Вынос общей логики рассчета фолдинга в отдельный класс
